requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
Click==7.0
coverage==4.5.4
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
importlib-metadata==0.19
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.1
pathlib2==2.3.4
pluggy==0.12.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.2
pytest==5.1.1
pytest-cov==2.7.1
six==1.12.0
wcwidth==0.1.7
Werkzeug==0.15.5
zipp==0.6.0
pycrypto
